🚚 move shared types from cf-core to types (#624)
