+++ modules/cf-core/src/index.ts
@@ -9,26 +9,6 @@ export { StateChannel } from "./models";
 export * from "./node";
 export * from "./private-keys-generator";
 export {
-  CreateChannelMessage,
-  DepositConfirmationMessage,
-  DepositStartedMessage,
-  DepositFailedMessage,
-  EventEmittedMessage,
-  InstallMessage,
-  InstallVirtualMessage,
-  NODE_EVENTS,
-  NodeMessageWrappedProtocolMessage,
-  ProposeMessage,
-  RejectInstallVirtualMessage,
-  RejectProposalMessage,
-  UninstallMessage,
-  UninstallVirtualMessage,
-  UpdateStateMessage,
-  WithdrawConfirmationMessage,
-  WithdrawFailedMessage,
-  WithdrawStartedMessage
-} from "./types";
-export {
   getCreate2MultisigAddress,
   bigNumberifyJson,
   deBigNumberifyJson
