+++ modules/cf-core/src/machine/enums.ts
@@ -1,3 +1,5 @@
+export { Protocol } from "../types";
+
 enum Opcode {
   /**
    * Called at the end of execution before the return value to store a commitment
@@ -26,16 +28,4 @@ enum Opcode {
   PERSIST_STATE_CHANNEL
 }
 
-enum Protocol {
-  Install = "install",
-  InstallVirtualApp = "install-virtual-app",
-  Setup = "setup",
-  Propose = "propose",
-  TakeAction = "takeAction",
-  Uninstall = "uninstall",
-  UninstallVirtualApp = "uninstall-virtual-app",
-  Update = "update",
-  Withdraw = "withdraw"
-}
-
-export { Opcode, Protocol };
+export { Opcode };
